01_etl/postgres_to_es/requirements.txt
python-dotenv==0.20.0
psycopg2-binary==2.9.3
requests==2.27.1
pydantic==1.9.1
